+++ test/worldwide/time_zone_test.rb
@@ -99,5 +99,10 @@ module Worldwide
 
       assert_equal "(GMT+12:00) Auckland, Wellington", zone.to_s
     end
+
+    test "#to_s display for Rails 8.1 zones America/Nuuk and America/Asuncion" do
+      assert_equal "(GMT-03:00) Greenland", Worldwide::TimeZone.new("America/Nuuk").to_s
+      assert_equal "(GMT-04:00) Asuncion", Worldwide::TimeZone.new("America/Asuncion").to_s
+    end
   end
 end